+++ RELEASE_NOTES.md
@@ -5,6 +5,8 @@ A new header is inserted each time a *tag* is created.
 
 ## v1.15.2 (currently main branch)
 
+- engine: add support for PCSS (Percentage Closer Soft Shadows)
+
 ## v1.15.1
 
 - engine: add support for DPCF (PCF shadows with contact hardening)